삭제 시 확인 다이얼로그 표시

+++ app/src/main/java/com/stopstone/musicplaylist/ui/detail/TrackDetailActivity.kt
@@ -1,6 +1,7 @@
 package com.stopstone.musicplaylist.ui.detail
 
 import android.app.Activity
+import android.content.DialogInterface
 import android.os.Bundle
 import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
@@ -8,6 +9,7 @@ import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.lifecycleScope
 import androidx.lifecycle.repeatOnLifecycle
 import androidx.navigation.navArgs
+import com.google.android.material.dialog.MaterialAlertDialogBuilder
 import com.stopstone.musicplaylist.databinding.ActivityTrackDetailBinding
 import com.stopstone.musicplaylist.ui.detail.viewmodel.TrackDetailViewModel
 import com.stopstone.musicplaylist.util.loadImage
@@ -77,7 +79,19 @@ class TrackDetailActivity : AppCompatActivity() {
         }
 
         binding.btnTrackDetailDelete.setOnClickListener {
-            viewModel.deleteTrack()
+            setDialogBuilder()
+        }
+    }
+
+    private fun setDialogBuilder() {
+        MaterialAlertDialogBuilder(this).apply {
+            setTitle("음악 삭제")
+            setMessage("정말로 삭제하시겠습니까?")
+            setNegativeButton("취소", null)
+            setPositiveButton("삭제") { dialogInterface: DialogInterface, i: Int ->
+                viewModel.deleteTrack()
+            }
+            show()
         }
     }
